targets: add workaround in mulitilib for x86_64
This is intentional, GCC does not pickup very well the options for
multilib.  This solve the issue with zlib while compiling the static
GCC; reported by some people trying to bootstrap the stage 0 under
other GNU/Linux distributions.

In this oportunity, the option --enable-long-long has been removed
for a more generic target.