qi: make some directory options to be uppercase
Previously, the -o option was duplicated with the option for .order files
(a bug).  All the options to be lowercase or uppercase cannot be possible,
since I try to maintain compatibility with the options of Graft.  So, all
the options which specified directories, except the target directory, are
uppercase now.