recipes: devel/binutils: compile it without target-specific directory
Following the example of other distros, including "Linux From Scratch":

Normally, the tooldir (the directory where the executables will
ultimately be located) is set to $(exec_prefix)/$(target_alias).
For example, x86_64 machines would expand that to
/usr/x86_64-unknown-linux-gnu. Because this is a custom system,
this target-specific directory in /usr is not required.
$(exec_prefix)/$(target_alias) would be used if the system was
used to cross-compile (for example, compiling a package on an
Intel machine that generates code that can be executed on PowerPC
machines).