sinit being replaced by sysvinit
OK.. sinit is great, but sysvinit is more complete in the sense of
commands (reboot, poweroff, shutdown) and its documentation.  It
also does not need to trick agetty as a supervised service, as we do
previously; in a quick experiment booting with sysvinit reveals less
seconds on the boot time, releasing between 1-2MB of RAM.

This is the release 2.89 of sysvinit, which I've contributed.  :-)