qi: do not set/change file mode bits using chmod(1)
Currently Qi eliminates the $srcdir if it exists previously.
This is a sane rule if you want to build or rebuild your source with a clean
source tree.  Setting sane permissions won't work from here, because the
build() must be called, then the unpack() to produce the $srcdir.

This does not make these permissions accommodated to set.