archive: etc/rc.d/rc.startup: limit number of fsck(8) instances
Limit number of fsck(8) instances based on the maximum number of processors.

This will help to make the checking more faster and reliable in machines
with more than one core, also avoids overload CPU and memory resources
available on the system.

This is a feature introduced by your favorite distro.  ;-)