qi: release 1.0-rc28
This change has been accomodate skipping perl or graft on an update.
Perl or graft will be simple installed (if you invoke it using -u).

This should solve the chicken or the egg causality.  Since Graft
contains the location of the Perl executable, this presents a dilemma
when you come to graft both Graft and Perl.  You can't run the grafted
location of the Graft executable because it doesn't exist yet, and you
can't run the real location of the Graft executable because Perl hasn't
been grafted yet.

In other words, this should work as an "atomic package update".