qi: reformulate, accommodate SIGINT cancelation on delete mode
Originally the idea is not to allow interruptions because if a package
(to be deleted) is canceled, it may leave unexpected results.  In this
commit, the lines that trap the signals have been accommodated giving
the possibility to cancel with for example, Ctrl+C when scanning for
conflicts.  This is uniform now with the install_mode().