qi: do not use tarlz for decompress .tar.lz files when unpack()
Previously, tarlz was introduced in the commit
c9063f925b03623920481c9cc41ae8462bdd0b87

The problem is that the .tar.lz has been produced using another archiver format,
like the one of GNU tar, tarlz won't be able to decompress the GNU format, because
simply does not understand this format.  So we revert the change in order to unpack
"global" .tar.lz files combining lzip and (GNU) tar.