Do not disable KMS by default in order to make use of the direct render capabilities
Let the kernel to set the video mode.  Previously when this was disabled the idea was to make use
of the "natural" resolution in the console, but this causes to disable the Direct Render
Capabilities in Xorg.  When 'nomodeset' is used the kernel does not try to load video drivers, it
tries to make uso of the BIOS modes instead until Xorg is loaded.