recipes: Rename 'pkghashtag' variable to 'pkgcategory'
Previously, and ideally, we tried to use the hash tag (#) to compose
the variable 'pkghashtag' that defines the category of a package.
Unfortunately, this was giving errors in the build process, since it
seems that some projects (and I suspect that the ash shell from
busybox-1.31.1) interpret it as embedded and therefore ignore it,
leaving the path to search incomplete.

Signed-off-by: Matías Fonzo <selk@dragora.org>