Turn on -D_FORTIFY_SOURCE=2 by default plus other security enhancements
Currently QICPPFLAGS is used to pass -D_FORTIFY_SOURCE=2 to the
CPPFLAGS that are used in recipes; unfortunately, there are sources
that do not even respect or include the CPPFLAGS, others use another
name for this flag.  Anyway, we make sure to always use this measure
through the compiler.

Also, in this opportunity we have applied patches for the compiler
(Thanks to "Alpine Linux") to improve security in Dragora by default.

Signed-off-by: Matías Fonzo <selk@dragora.org>