Replace socklog with sysklogd, logrotate alternatives
Back in time, the first Dragora series included sysklogd and logrotate.
Later it was replaced by Runit; already in Dragora 3 the same scheme
for sysklogd and logrotate was tried to be followed but in communication
with Wayne Marshall[1] the course was changed.  This is also partly
because sysklogd was not being maintained for some time and a reliable
alternative was sought, but now sysklogd is being maintained, improved.
It also includes the klog functionality inside, so it is no longer
necessary to have "klogd" as a separate service running.

[1] https://lists.nongnu.org/archive/html/dragora-users/2017-05/msg00028.html

Signed-off-by: Matías Fonzo <selk@dragora.org>