Start mass upgrade
A mass upgrade is needed to continue with the current components and
the components or software that must be added.  This bunch of updates
also includes the following relevant changes:

- Binutils has been reverted to the version 2.35.2, plus patched.
- GCC has been reverted to the release 10 from latest snapshot.

After testing binutils 2.36+ and GCC 11+ it gives errors.  Errors when
there are threads, random errors in the build (bootstrap) and errors for
the build of binutils when you want to process libctf as dynamic.

I have no plans, after a long burnout to try those versions again, since
the current ones (binutils 2.35.2+ and GCC 10+) are working very well in
Dragora.

Continuing with the changes "gettext-tiny" has been replaced by GNU gettext,
which is a bigger and more complex software, but I think necessary to cover
a complete NLS (Native Language Support) base.  There are also other minor
but not less important changes and fixes in some recipes.

Signed-off-by: Matías Fonzo <selk@dragora.org>