recipes: Silent stderr on post-install for the installer
The Dragora installer will report on missing symbols when Qi run the
post-install instructions that come from packages related to the X
environment, where part of this is not available.  Manipulating order
on the installer would be a very large task, so it is preferable to
ignore in case of error.

Signed-off-by: Matías Fonzo <selk@dragora.org>