Do not recommend the force option to bootstrap for first time
The content of the CHEATSHEET.md file must be moved to the dragora-handbook:

Maybe an extra note about whether the -f (force) option may be added to the
manual (dragora-handbook) in order to indicate whether the package directory
is populated by packages for some reason: it was obtained from a repository
or the user is using a CD provided with the sources.  For these cases,
suggest the -f (force) option when bootstrapping for first time.

Signed-off-by: Matías Fonzo <selk@dragora.org>